+++ src/main/java/nl/stil4m/mollie/domain/MethodAmount.java
@@ -1,25 +1,27 @@
 package nl.stil4m.mollie.domain;
 
+import java.math.BigDecimal;
+
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class MethodAmount {
 
-    private final Double minimum;
-    private final Double maximum;
+    private final BigDecimal minimum;
+    private final BigDecimal maximum;
 
     @JsonCreator
-    public MethodAmount(@JsonProperty("minimum") Double minimum,
-                        @JsonProperty("maximum") Double maximum) {
+    public MethodAmount(@JsonProperty("minimum") BigDecimal minimum,
+                        @JsonProperty("maximum") BigDecimal maximum) {
         this.minimum = minimum;
         this.maximum = maximum;
     }
 
-    public Double getMinimum() {
+    public BigDecimal getMinimum() {
         return minimum;
     }
 
-    public Double getMaximum() {
+    public BigDecimal getMaximum() {
         return maximum;
     }
 }
